refactor(03-06): reorg portfolio_handler into subdomain packages (D-11)
- git mv position.py + position_manager.py -> position/
- git mv transaction.py + transaction_manager.py -> transaction/
- git mv cash_manager.py -> cash/; metrics_manager.py -> metrics/
- add per-package __init__.py re-exporting public classes (+ TransactionType/PositionSide from core.enums)
- rewire portfolio.py manager imports + 4 test files to new subdomain paths
- pure history-preserving move; suite 321 pass / collected 326 unchanged; mypy --strict clean; behavioral oracle byte-exact
