+++ itrader/portfolio_handler/metrics/__init__.py
@@ -1,0 +1,10 @@
+"""
+Metrics subdomain package.
+
+Re-exports the public metrics manager + its snapshot entities so consumer import
+paths stay short after the D-11 subdomain reorg (pure move, no behavior change).
+"""
+
+from .metrics_manager import MetricsManager, PortfolioSnapshot, PerformanceMetrics
+
+__all__ = ["MetricsManager", "PortfolioSnapshot", "PerformanceMetrics"]
