+++ itrader/portfolio_handler/position/__init__.py
@@ -1,0 +1,12 @@
+"""
+Position subdomain package.
+
+Re-exports the public position entity + manager so consumer import paths stay
+short after the D-11 subdomain reorg (pure move, no behavior change).
+"""
+
+from itrader.core.enums import PositionSide
+from .position import Position
+from .position_manager import PositionManager
+
+__all__ = ["Position", "PositionSide", "PositionManager"]
