+++ itrader/portfolio_handler/transaction/__init__.py
@@ -1,0 +1,12 @@
+"""
+Transaction subdomain package.
+
+Re-exports the public transaction entity + manager so consumer import paths stay
+short after the D-11 subdomain reorg (pure move, no behavior change).
+"""
+
+from itrader.core.enums import TransactionType
+from .transaction import Transaction
+from .transaction_manager import TransactionManager
+
+__all__ = ["Transaction", "TransactionType", "TransactionManager"]
